backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Security
slowapi==0.1.9
authlib==1.3.0
pyotp==2.9.0
qrcode[pil]==7.4.2

# Document parsing
pdfplumber==0.11.4
python-docx==1.1.2

# OCR for image-based statements
pytesseract==0.3.13
Pillow==11.1.0

# Task queue
celery[redis]==5.4.0
redis==5.2.1

# AI integration
httpx==0.28.1

# Import support
openpyxl==3.1.5

# Payments
stripe>=8.0.0

# Utilities
python-dateutil==2.9.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
aiosqlite==0.20.0
